Update details component and rollerplace form

+++ src/components/RollerPlaceDetails.vue
@@ -1,15 +1,32 @@
 <template>
-  <div>
-    <img :src="place.image" :alt="place.name" />
-    <p>Nombre: {{ place.name }}</p>
-    <p>Descripción: {{ place.description }}</p>
-    <p>Dificultas: {{ place.level }}</p>
-    <p>Ciudad: {{ place.city }}</p>
-    <p>slalom: ✅</p>
-  </div>
-  <v-btn @click="updatePlaceClicked()">editar</v-btn>
-  <v-btn @click="deletePlaceClicked()">borrar</v-btn>
-  <div>mapa</div>
+  <v-row>
+    <v-col cols="6">
+      <v-card class="mx-auto">
+        <v-img class="align-end text-white" height="200" :src="place.image" cover> </v-img>
+        <v-card-text>
+          <v-card-title class="name"> {{ place.name }} </v-card-title>
+
+          <div>Descripción: {{ place.description }}</div>
+          <div>Dificultad: {{ place.level }}</div>
+          <div>Ciudad: {{ place.city }}</div>
+          <div>Slalom: {{ place.slalom }}</div>
+        </v-card-text>
+
+        <v-card-actions v-if="isAuthor">
+          <v-btn @click="updatePlaceClicked()" color="purple"> Editar </v-btn>
+          <v-btn @click="deletePlaceClicked()" color="red"> Borrar </v-btn>
+        </v-card-actions>
+      </v-card>
+    </v-col>
+    <v-col cols="6">
+      <mapbox-map :accessToken="mapboxToken" :center="place.location" :zoom="15" mapStyle="streets-v11">
+        <mapbox-navigation-control position="top-right" />
+        <mapbox-geolocate-control />
+
+        <mapbox-marker :lngLat="place.location" />
+      </mapbox-map>
+    </v-col>
+  </v-row>
 </template>
 
 <script lang="ts">
@@ -23,9 +40,15 @@ export default defineComponent({
     const store = useRollerMapStore();
     return { store };
   },
+  computed: {
+    isAuthor() {
+      return this.store.user._id === this.place.author;
+    },
+  },
   data() {
     return {
       place: {} as RollerPlace,
+      mapboxToken: "pk.eyJ1IjoiYWVzY29sYW5vIiwiYSI6ImNsMTgzcThjajFhNzAzaXNnbXJicmRwaTcifQ.8thc3vNfqF8E7yHLqap9PQ",
     };
   },
   props: {
@@ -40,6 +63,7 @@ export default defineComponent({
       this.place = placeFound;
     }
   },
+
   methods: {
     updatePlaceClicked() {
       this.$router.push({
@@ -48,10 +72,12 @@ export default defineComponent({
       });
     },
     deletePlaceClicked() {
-      try {
-        this.store.deleteRollerPlace(this.place._id);
-      } catch (error) {
-        return error;
+      if (this.isAuthor && this.place._id) {
+        try {
+          this.store.deleteRollerPlace(this.place._id);
+        } catch (error) {
+          return error;
+        }
       }
     },
   },
